Move former __init__.py code to util.py
Creates slightly more aesthetic import statements such as
`import manubot.cite.util` rather than `import manubot.cite.cite`.
