+++ tests/test_imports.py
@@ -1,10 +1,10 @@
 
 def test_imports():
     import manubot.cite
-    import manubot.cite.cite
+    import manubot.cite.util
     import manubot.cite.arxiv
     import manubot.cite.doi
     import manubot.cite.pubmed
     import manubot.cite.url
     import manubot.process.manuscript
-    assert isinstance(manubot.cite.cite.citeproc_retrievers, dict)
+    assert isinstance(manubot.cite.util.citeproc_retrievers, dict)
